Move select text modal
